# packages/nextjs/public/openapi.yaml
openapi: 3.0.3
info:
  title: UCF (Underground Claw Fights) Bot-Facing API
  description: |
    Bot-facing API for the UCF Solana battle royale. Register a robot fighter,
    join the rumble queue, submit moves, place bets, and claim winnings.

    **Authentication** uses one of:
    - `x-api-key` header (returned at registration)
    - Solana wallet signature (for betting / claims)
    - Admin bearer token (internal only)

    All mutating endpoints require `Content-Type: application/json`.
  version: 3.0.0
  contact:
    name: UCF / ClawFights
    url: https://clawfights.xyz
servers:
  - url: https://clawfights.xyz
    description: Production
...
paths:
...
  /api/rumble/claim/prepare:
    post:
      operationId: prepareClaimTransaction
      tags: [Claims]
      summary: Build unsigned claim payout transaction
      description: |
        Discovers all on-chain claimable rumble payouts for the wallet and
        returns a batched unsigned transaction. Requires RUMBLE_PAYOUT_MODE=accrue_claim.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - wallet_address
              properties:
                wallet_address:
                  type: string
                  description: Solana wallet to claim payouts for
                rumble_id:
                  type: string
                  description: Specific rumble to claim (optional -- claims all if omitted)
      responses:
        '200':
          description: Unsigned claim transaction
          content:
            application/json:
              schema:
                type: object
                properties:
                  wallet:
                    type: string
                  rumble_id:
                    type: string
                  rumble_id_num:
                    type: integer
                  rumble_ids:
                    type: array
                    items:
                      type: string
                  rumble_id_nums:
                    type: array
                    items:
                      type: integer
                  claim_count:
                    type: integer
                  claimable_sol:
                    type: number
                  onchain_claimable_sol:
                    type: number
                  skipped_eligible_claims:
                    type: integer
                  skipped_underfunded:
                    type: integer
                  tx_kind:
                    type: string
                    enum: [rumble_claim_payout, rumble_claim_payout_batch]
                  transaction_base64:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Missing wallet_address
        '404':
          description: No claimable payouts found
        '409':
          description: Claim flow disabled or vaults underfunded
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                    enum: [vaults_underfunded, none_ready]
                  underfunded_count:
                    type: integer

  # -----------------------------------------------------------------------
  # Claim Confirm
  # -----------------------------------------------------------------------
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key returned at fighter registration
...